+++ Chartboost/build.gradle
@@ -1,4 +1,4 @@
-project.version = '8.2.0.0'
+project.version = '8.2.0.1'
 project.ext.networkName = 'chartboost'
 
 apply from: '../shared-build.gradle'
