+++ Chartboost/src/main/java/com/mopub/mobileads/ChartboostAdapterConfiguration.java
@@ -8,17 +8,21 @@ import androidx.annotation.Nullable;
 
 import com.chartboost.sdk.Chartboost;
 import com.chartboost.sdk.Libraries.CBLogging;
+import com.chartboost.sdk.Privacy.model.GDPR;
 import com.mopub.common.BaseAdapterConfiguration;
 import com.mopub.common.MoPub;
 import com.mopub.common.OnNetworkInitializationFinishedListener;
 import com.mopub.common.Preconditions;
 import com.mopub.common.logging.MoPubLog;
+import com.mopub.common.privacy.ConsentStatus;
+import com.mopub.common.privacy.PersonalInfoManager;
 import com.mopub.mobileads.chartboost.BuildConfig;
 
 import java.util.Map;
 
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.CUSTOM;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.CUSTOM_WITH_THROWABLE;
+import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.LOAD_FAILED;
 
 public class ChartboostAdapterConfiguration extends BaseAdapterConfiguration {
 
@@ -28,6 +32,12 @@ public class ChartboostAdapterConfiguration extends BaseAdapterConfiguration {
     // Chartboost's keys
     private static final String APP_ID_KEY = "appId";
     private static final String APP_SIGNATURE_KEY = "appSignature";
+
+    // Chartboost specific ids
+    @Nullable
+    private static String mAppId;
+    @Nullable
+    private static String mAppSignature;
 
     // Adapter's keys
     private static final String ADAPTER_NAME = ChartboostAdapterConfiguration.class.getSimpleName();
@@ -73,40 +83,21 @@ public class ChartboostAdapterConfiguration extends BaseAdapterConfiguration {
         Preconditions.checkNotNull(listener);
 
         boolean networkInitializationSucceeded = false;
-
         synchronized (ChartboostAdapterConfiguration.class) {
             try {
                 if (configuration != null && !configuration.isEmpty()) {
-
-                    ChartboostShared.initializeSdk(context, configuration);
-
-                    final String appId = configuration.get(APP_ID_KEY);
-                    final String appSignature = configuration.get(APP_SIGNATURE_KEY);
-
-                    if (TextUtils.isEmpty(appId) || TextUtils.isEmpty(appSignature)) {
-                        MoPubLog.log(CUSTOM, ADAPTER_NAME, "Chartboost's initialization " +
-                                "succeeded, but unable to call Chartboost's startWithAppId(). " +
-                                "Ensure Chartboost's " + APP_ID_KEY + " and " + APP_SIGNATURE_KEY +
-                                "are populated on the MoPub dashboard. Note that initialization on " +
-                                "the first app launch is a no-op.");
-                    } else {
-                        Chartboost.startWithAppId(context, appId, appSignature);
-                    }
-
-                    Chartboost.setMediation(Chartboost.CBMediation.CBMediationMoPub,
-                            MoPub.SDK_VERSION, BuildConfig.VERSION_NAME);
-                    Chartboost.setDelegate(sDelegate);
-                    Chartboost.setAutoCacheAds(false);
-                    networkInitializationSucceeded = true;
+                    networkInitializationSucceeded = initializeChartboostSdk(context, configuration);
                 } else {
                     MoPubLog.log(CUSTOM, ADAPTER_NAME, "Chartboost's initialization via " +
-                            ADAPTER_NAME + " not started as the context calling it is missing or null.");
+                            ADAPTER_NAME + " not started as the context calling it, or configuration info is missing or null.");
                 }
-            } catch (Exception e) {
-                MoPubLog.log(CUSTOM_WITH_THROWABLE, "Initializing Chartboost has encountered " +
-                        "an exception.", e);
+            } catch (Exception exception) {
+                MoPubLog.log(CUSTOM_WITH_THROWABLE, ADAPTER_NAME, "Initializing Chartboost has encountered " +
+                        "an exception.", exception.getMessage());
             }
         }
+
+        setChartboostLogLevel();
 
         if (networkInitializationSucceeded) {
             listener.onNetworkInitializationFinished(ChartboostAdapterConfiguration.class,
@@ -115,10 +106,95 @@ public class ChartboostAdapterConfiguration extends BaseAdapterConfiguration {
             listener.onNetworkInitializationFinished(ChartboostAdapterConfiguration.class,
                     MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
         }
+    }
 
+    /**
+     * Initialize the Chartboost SDK for the provided application id and app signature.
+     */
+    public static synchronized boolean initializeChartboostSdk(@NonNull Context context,
+                                                               @NonNull Map<String, String> configuration) {
+        Preconditions.checkNotNull(context);
+        Preconditions.checkNotNull(configuration);
+
+        if(Chartboost.isSdkStarted())
+            return true;
+
+        // Pass the user consent from the MoPub SDK to Chartboost as per GDPR
+        PersonalInfoManager personalInfoManager = MoPub.getPersonalInformationManager();
+
+        final boolean canCollectPersonalInfo = MoPub.canCollectPersonalInformation();
+        final boolean shouldAllowLegitimateInterest = MoPub.shouldAllowLegitimateInterest();
+
+        if (personalInfoManager != null && personalInfoManager.gdprApplies() == Boolean.TRUE) {
+            if (shouldAllowLegitimateInterest) {
+                boolean isExplicitNoConsent = personalInfoManager.getPersonalInfoConsentStatus() == ConsentStatus.EXPLICIT_NO
+                        || personalInfoManager.getPersonalInfoConsentStatus() == ConsentStatus.DNT;
+                addChartboostPrivacyConsent(context, !isExplicitNoConsent);
+            } else {
+                addChartboostPrivacyConsent(context, canCollectPersonalInfo);
+            }
+        }
+
+        // Validate Chartboost args
+        if (!configuration.containsKey(APP_ID_KEY)) {
+            MoPubLog.log(LOAD_FAILED, ADAPTER_NAME,
+                    MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR.getIntCode(),
+                    MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
+
+            throw new IllegalStateException("Chartboost rewarded video initialization" +
+                    " failed due to missing application ID.");
+        }
+
+        if (!configuration.containsKey(APP_SIGNATURE_KEY)) {
+            MoPubLog.log(LOAD_FAILED, ADAPTER_NAME,
+                    MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR.getIntCode(),
+                    MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
+
+            throw new IllegalStateException("Chartboost rewarded video initialization" +
+                    " failed due to missing application signature.");
+        }
+
+        final String appId = configuration.get(APP_ID_KEY);
+        final String appSignature = configuration.get(APP_SIGNATURE_KEY);
+
+        if (!TextUtils.isEmpty(appId) && !TextUtils.isEmpty(appSignature)) {
+            if (appId.equals(mAppId) && appSignature.equals(mAppSignature)) {
+                // We don't need to reinitialize.
+                return false;
+            }
+        }
+
+        mAppId = appId;
+        mAppSignature = appSignature;
+
+        // Perform all the common SDK initialization steps including startAppWithId
+        Chartboost.startWithAppId(context, mAppId, mAppSignature);
+        Chartboost.setMediation(Chartboost.CBMediation.CBMediationMoPub, MoPub.SDK_VERSION,
+                new ChartboostAdapterConfiguration().getAdapterVersion());
+        Chartboost.setDelegate(sDelegate);
+        Chartboost.setAutoCacheAds(false);
+        return true;
+    }
+
+    private static void addChartboostPrivacyConsent(Context context, boolean canCollectPersonalInfo) {
+        if (context == null) {
+            MoPubLog.log(CUSTOM, ADAPTER_NAME, "Skipped setting Chartboost Privacy consent " +
+                    "as context is null.");
+            return;
+        }
+
+        if (canCollectPersonalInfo) {
+            MoPubLog.log(CUSTOM, ADAPTER_NAME, "Setting Chartboost GDPR data use consent as BEHAVIORAL");
+            Chartboost.addDataUseConsent(context, new GDPR(GDPR.GDPR_CONSENT.BEHAVIORAL));
+        } else {
+            MoPubLog.log(CUSTOM, ADAPTER_NAME, "Setting Chartboost GDPR data use consent as NON_BEHAVIORAL");
+            Chartboost.addDataUseConsent(context, new GDPR(GDPR.GDPR_CONSENT.NON_BEHAVIORAL));
+        }
+    }
+
+    private void setChartboostLogLevel() {
         MoPubLog.LogLevel mopubLogLevel = MoPubLog.getLogLevel();
         CBLogging.Level chartboostLogLevel = getChartboostLogLevel(mopubLogLevel);
-
         Chartboost.setLoggingLevel(chartboostLogLevel);
     }
